feat: add component SearchBar
