--- FRW/Modules/NFT/Views/SearchBar.swift
@@ -1,70 +1,0 @@
-//
-//  SearchBar.swift
-//  Flow Wallet
-//
-//  Created by cat on 2022/7/6.
-//
-
-import SwiftUI
-
-// MARK: - SearchBar
-
-struct SearchBar: View {
-    @Binding
-    var text: String
-    @State
-    private var isEditing = false
-    var purpose: String = "Search..."
-    var iconColor: Color = .gray
-
-    var body: some View {
-        HStack {
-            TextField(purpose, text: $text)
-                .padding(8)
-                .padding(.horizontal, 25)
-                .background(Color(.systemGray6))
-                .cornerRadius(8)
-                .overlay {
-                    HStack {
-                        Image(systemName: "magnifyingglass")
-                            .foregroundColor(iconColor)
-                            .frame(minWidth: 0, maxWidth: .infinity, alignment: .leading)
-                            .padding(.leading, 8)
-                    }
-                }
-                .animation(.easeInOut, value: self.isEditing)
-                .onTapGesture {
-                    self.isEditing = true
-                }
-
-            if isEditing {
-                Button {
-                    self.isEditing = false
-                    self.text = ""
-                    UIApplication.shared.sendAction(
-                        #selector(UIResponder.resignFirstResponder),
-                        to: nil,
-                        from: nil,
-                        for: nil
-                    )
-
-                } label: {
-                    Image(systemName: "multiply.circle.fill")
-                        .foregroundColor(.gray)
-                        .padding(.trailing, 8)
-                }
-                .padding(.trailing, 10)
-                .transition(.move(edge: .trailing))
-                .animation(.easeInOut, value: self.isEditing)
-            }
-        }
-    }
-}
-
-// MARK: - SearchBar_Previews
-
-struct SearchBar_Previews: PreviewProvider {
-    static var previews: some View {
-        SearchBar(text: .constant(""))
-    }
-}
